Creating EmployeeServiceImpl implementation of creating new employee

+++ employee-details-service/src/main/java/dev/kilima/sudaksha/task/employee/repository/EmployeeDetailsRepository.java
@@ -10,7 +10,7 @@ import dev.kilima.sudaksha.task.employee.entity.EmployeeDetails;
 @Repository
 public interface EmployeeDetailsRepository extends JpaRepository<EmployeeDetails, String>{
 	@Query(value = "select max(emp.empid) from EmployeeDetails emp")
-	Optional<Long> getMaxId();
+	Optional<Integer> getMaxId();
 
 	@Query(value = "select emp.availability from EmployeeDetails emp where emp.empid=?1")
 	String checkAvailability(String empid);
