+++ employee-details-service/src/main/java/dev/kilima/sudaksha/task/employee/service/EmployeeService.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 import dev.kilima.sudaksha.task.employee.entity.EmployeeDetails;
 
 public interface EmployeeService {
-	Optional<Long> getMaxId();
+	Optional<Integer> getMaxId();
 	Optional<EmployeeDetails> checkEmployeeAvailability(String empid);
 	String createNewEmployee(EmployeeDetails emp);
 }
